requirements.txt
# Python dependencies for LLM Code Generation Thesis

# API Clients for LLMs
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0

# Code Quality Analysis
radon>=5.1.0
pylint>=3.0.0
flake8>=6.1.0
lizard>=1.17.10
mccabe>=0.7.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Testing & Execution
pytest>=7.4.0
pytest-timeout>=2.1.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0

# Jupyter (for analysis)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Optional: Docker SDK (for code sandbox)
docker>=6.1.0

# Type hints
mypy>=1.5.0
types-requests>=2.31.0
